requirements.txt
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient>=2.2.5
python-dotenv==1.0.1
pytest>=7.4.0
pytest-django>=4.7.0
requests>=2.31.0 
